Vérification de l'intégrité de la base de données

+++ src/Vue/FormulaireAdmin/SupprimerAvion.java
@@ -16,6 +16,7 @@ public class SupprimerAvion extends JPanel {
     private JComboBox comboBoxAvion;
     private JButton supprimerAvionButton;
     private JPanel panelSupAvion;
+    private JLabel labelError;
     private TableauVols tableauVols;
     public SupprimerAvion() {
 
@@ -23,12 +24,31 @@ public class SupprimerAvion extends JPanel {
             @Override
             public void actionPerformed(ActionEvent e) {
                 Avion avion = (Avion)comboBoxAvion.getSelectedItem();
-                comboBoxAvion.removeItem(avion);
-                tableauVols.deleteAvion(avion);
+                boolean exist = false;//Vérification de l'intégrité de la base de données
+                for (int i =0; i< tableauVols.getTableauDeVols().size();i++){
+                    if(tableauVols.getTableauDeVols().get(i).getAvion().getRef().equals(avion.getRef())){
+                        exist = true;
+                    }
+                }
+                if(!exist){
+                    comboBoxAvion.removeItem(avion);
+                    tableauVols.deleteAvion(avion);
+                    labelError.setText("Supprimer");
+                }else {
+                    labelError.setText("Avion utiliser dans un vol");
+                }
+
+            }
+        });
+        comboBoxAvion.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                labelError.setText("");
             }
         });
     }
     public void updateBox() {
+        labelError.setText("");
         tableauVols = new TableauVols("Suppresion Type Avion");
         ArrayList<Avion> tabTypeAvion = tableauVols.getTableauAvion();
         DefaultComboBoxModel def = new DefaultComboBoxModel();
