+++ src/Vue/FormulaireAdmin/SupprimerTypeAvion.java
@@ -16,22 +16,48 @@ public class SupprimerTypeAvion extends JPanel {
     private JComboBox comboBoxTypeAvion;
     private JButton supprimerTypeAvionButton;
     private JPanel panelSuppresionTypeAvion;
+    private JLabel labelError;
     private TableauVols tableauVols;
     public SupprimerTypeAvion() {
         supprimerTypeAvionButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                TypeAvion typeAvion = (TypeAvion) comboBoxTypeAvion.getSelectedItem();
-                comboBoxTypeAvion.removeItem(typeAvion);
-                tableauVols.deleteTypeAvion(typeAvion);
-
+                if(comboBoxTypeAvion.getSelectedItem() !=null){
+                    TypeAvion typeAvion = (TypeAvion) comboBoxTypeAvion.getSelectedItem();
+                    boolean exist = false;//Vérification de l'intégrité de la base de données
+                    for(int i =0 ; i<tableauVols.getTableauAvion().size();i++){
+                        if(tableauVols.getTableauAvion().get(i).getTypeAvion().getNom().equals(typeAvion.getNom())){
+                            exist = true;
+                        }
+                    }
+                    for(int j=0;j<tableauVols.getTableauMembreEquipage().size();j++){
+                        if(tableauVols.getTableauMembreEquipage().get(j).isQualified(typeAvion)){
+                            System.out.println("Qualified  ->" +tableauVols.getTableauMembreEquipage().get(j).getNom());
+                            exist = true;
+                        }
+                    }
+                    if(!exist){
+                        comboBoxTypeAvion.removeItem(typeAvion);
+                        tableauVols.deleteTypeAvion(typeAvion);
+                        labelError.setText("Supprimer");
+                    }else{
+                        labelError.setText("Type Avion associé à un avion ou un membre d'équipage est qualifier sur cet avion");//Supprimer avion pour supprimer type avion
+                    }
+                }
             }
         });
 
 
+        supprimerTypeAvionButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                labelError.setText("");
+            }
+        });
     }
 
-    public void updateBox() {
+    public void update() {
+        labelError.setText("");
         tableauVols = new TableauVols("Suppresion Type Avion");
         ArrayList<TypeAvion> tabTypeAvion = tableauVols.getTablaeuTypeAvion();
         DefaultComboBoxModel def = new DefaultComboBoxModel();
